Image support has been added to the applicaton. Image files are loaded onto Firebase by creating a directory with the corresponding post name, and then storing the image files into that directory.
The storage is read from in the BikeFeedViewController. The getData() API function from Firebase works in the background. After all of the posting objects are created then the photos start being downloaded into their respecive postings.

Two metadata values were added to the postings object, one stores the ID of the document corresponsing to the posting, and the other is an array which stores the ID's of the images. This is for obtaining the correct image file in the storage.

The timing issue was fixed by changing the objects in ViewListingViewController. In the old system there was a global image array that would be loaded when the ViewListingViewController was opened. Instead, I changed it so that the array of the actual posting is what the user interacts with. So as images are being loaded the user will be able to interact with them.

This is useful because it will not stop the user from interacting with the application while images are loading. So if someone wants to open up a listing that has lower quality images, they will not have to wait for all of the high quality ones in the other postings to load. This will be very noticeable when there are a lot of posts.

The only issue, is that the user is not notified when all of the pictures are loaded or if another is loaded. So they have to keep swiping to find out if there is another image.